fix number size calculation

+++ bit_utils.h
@@ -4,7 +4,7 @@
 #include "err.h"
 #include "bool.h"
 
-bool number_fits_in_bits(int number, int number_of_bits);
+bool number_fits_in_bits(long number, int number_of_bits);
 void write_binary_stream_to_buffer(int num, int number_of_bits, unsigned char* buffer);
